fix(kit): record every settled tool result before a turn ends
A model can emit several change_apply calls in one reply. The loop executed
all of them but returned on the FIRST apply result, so later results were
never recorded in history. The next request then carried an assistant tool
call with no matching tool result, and providers rejected it with "Tool
result is missing for tool call ..." - seen in the wild: one reply carried
two change_apply calls (draft-2, draft-3) and only the first was recorded,
so the resumed session tripped the provider on every request.

The apply terminal now records every settled result before the turn ends.
Draft/apply/review loop tests moved to review-lifecycle.test.ts with a
shared harness in loop/_shared/test-harness.ts, keeping both files under the
line cap.
